Backend/requirements.txt
openai-agents
requests
python-dotenv
fastapi
uvicorn